requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pandas==1.5.3
pika==1.1.0
azure-functions
azure-identity
azure-storage-blob
